split out Item class to own module
